added utility fuction addmarkers and comments for overview
